Add csrf protection in delete course

+++ openeclass-2.3/modules/course_info/delete_course.php
@@ -28,11 +28,12 @@ $require_current_course = TRUE;
 $require_prof = true;
 include '../../include/baseTheme.php';
 
+$token = $_SESSION['token'];
 $nameTools = $langDelCourse;
 $tool_content = "";
 
 if($is_adminOfCourse) {
-	if(isset($delete)) {
+	if(isset($delete) && ($token==$_GET['token'])) {
 		mysql_select_db("$mysqlMainDb",$db);
 		mysql_query("DROP DATABASE `$currentCourseID`");
 		mysql_query("DELETE FROM `$mysqlMainDb`.cours WHERE code='$currentCourseID'");
@@ -61,7 +62,7 @@ if($is_adminOfCourse) {
 		</tr>
 		<tr>
 		<th rowspan='2' class='left' width='220'>$langConfirmDel :</th>
-		<td width='52' align='center'><a href=\"".$_SERVER['PHP_SELF']."?delete=yes\">$langYes</a></td>
+		<td width='52' align='center'><a href=\"".$_SERVER['PHP_SELF']."?delete=yes&token=$token\">$langYes</a></td>
 		<td><small>$langByDel</small></td>
 		</tr>
 		<tr>
@@ -70,7 +71,7 @@ if($is_adminOfCourse) {
 		</tr>
 		</tbody>
 		</table>";
-		
+
 		$tool_content .= "<p align=\"right\"><a href=\"infocours.php\">$langBack</a></p>
 		</ul>
 		</div>";
